FWU: Remove the capsule acceptance authentication
We were initiallyplanning on making acceptance capsules authenticated.
However, the capsule it self has no payload so adding authentication is
not doable unless we change the acceptance capsule format.
On top of that the only real threat is being able to downgrade to an
older firmware. But the spec itself describes how to subvert similar
attacks by using rollback counters. So remove this requirement

Signed-off-by: Ilias Apalodimas <ilias.apalodimas@linaro.org>