FWU:  Describe an alternative acceptance path via EFI variables
When the spec was initially drafted UEFI implementations did not support
SetVariable at runtime on embedded boards. As a result an empty capsule
with a special GUID was created for the acceptance and revert cases.

Recently SetVariable at Runtime support has been added on open source
UEFI implementations and notably U-Boot. So let's define an alternative
update and downgrade path, by using a new EFI variable called ABAction.
On top of that define ABStatus, so the firmware can comunicate to the
OS the supported upstate status, since ESRT cannot change at runtime.

Define the GUID and 2 bits for ABAction
bit0: Reject the upgrade and revert to the previous bank
bit1: Accept the upgrade configure the new active banks

and describe an upodated version of the ESRT for ABStatus

Signed-off-by: Ilias Apalodimas <ilias.apalodimas@linaro.org>