FWU: Describe an alternative acceptance path via
When the spec was initially drafted UEFI implementations did not support
SetVariable at runtime on embedded boards. As a result an empty capsule
with a special GUID was created for the acceptance and revert cases.

Recently SetVariable at Runtime support has been added on open source
UEFI implementations and notably U-Boot. So let's define an alternative
update and downgrade path, by using a new EFI variable called ABAction.
On top of that define ABLastAttemptStatus, so the firmware can
communicate to the OS the current state.

Signed-off-by: Ilias Apalodimas <ilias.apalodimas@linaro.org>